Split autonomy constants.
